generator/requirements.txt
bleach==3.1.4
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cryptography==2.9
cwlgen-kclhi==0.4.0
docutils==0.16
gitdb==4.0.2
h11==0.9.0
httptools==0.1.1
idna==2.9
jeepney==0.4.3
keyring==21.2.0
nodeenv==1.3.5
nose==1.3.7
pkginfo==1.5.0.1
pycparser==2.20
Pygments==2.6.1
PyYAML==5.3.1
readme-renderer==25.0
requests==2.23.0
requests-toolbelt==0.9.1
ruamel.yaml==0.16.5
SecretStorage==3.1.2
six==1.14.0
smmap==3.0.1
starlette==0.13.2
tqdm==4.45.0
twine==3.1.1
urllib3==1.25.8
uvicorn==0.11.3
uvloop==0.14.0
webencodings==0.5.1
websockets==8.1
